+++ graph/SCC.cpp
@@ -2,30 +2,30 @@
 #include "../template/template.cpp"
 
 struct SCC{
-	vector<vector<int>>G,rG;
+	vector<vector<ll>>G,rG;
 	vector<bool>used;
-	vector<int>vs;
-	void addEdge(int x,int y){
+	vector<ll>vs;
+	void addEdge(ll x,ll y){
 		G[x].push_back(y);
 		rG[y].push_back(x);
 	}
-	void dfs(int x){
+	void dfs(ll x){
 		used[x]=true;
-		for(int i:G[x])if(!used[i])dfs(i);
+		for(ll i:G[x])if(!used[i])dfs(i);
 		vs.push_back(x);
 	}
-	vector<int>cmp;
-	void rdfs(int v,int k){
+	vector<ll>cmp;
+	void rdfs(ll v,ll k){
 		used[v]=true;
 		cmp[v]=k;
-		for(int i:rG[v])if(!used[i])rdfs(i,k);
+		for(ll i:rG[v])if(!used[i])rdfs(i,k);
 	}
 	void scc(){
 		fill(all(used),false);
 		rep(i,len(G))if(!used[i])dfs(i);
 		fill(all(used),false);
-		int k=0;
+		ll k=0;
 		rev(i,len(vs))if(!used[vs[i]])rdfs(vs[i],k++);
 	}
-	SCC(int N):G(N),rG(N),used(N),cmp(N){}
+	SCC(ll N):G(N),rG(N),used(N),cmp(N){}
 };
